+++ test/property_damage/replay_test.exs
@@ -52,7 +52,7 @@ defmodule PropertyDamage.ReplayTest do
     def teardown(_context), do: :ok
 
     @impl true
-    def execute(%Bump{}, _context), do: {:ok, [%Counted{amount: 1}]}
+    def execute(%Bump{}, _context, _runtime), do: {:ok, [%Counted{amount: 1}]}
   end
 
   defmodule CounterModel do